fcl: Split fclmanagercache.h into header and implementation.
